// v1/randstr/src/types.ts
export interface IAppProps {
}

export interface IAppState {
    count: number;
    length: number;

    useLowers: boolean;
    useUppers: boolean;
    useDigits: boolean;
    useSymbols: boolean;
    useExclamationMark: boolean;
    useDoubleQuotationMark: boolean;
    useNumberSign: boolean;
    useDollarSign: boolean;
    usePercentSign: boolean;
    useAmpersandSign: boolean;
    useApostrophe: boolean;
    useLeftParenthesis: boolean;
    useRightParenthesis: boolean;
    useAsterisk: boolean;
    usePlusSign: boolean;
    useComma: boolean;
    useMinusSign: boolean;
    usePeriod: boolean;
    useSlash: boolean;
    useColon: boolean;
    useSemicolon: boolean;
    useLessThanSign: boolean;
    useEqualSign: boolean;
    useGreaterThanSign: boolean;
    useQuestionMark: boolean;
    useCommercialAtSign: boolean;
    useLeftSquareBracket: boolean;
    useBackslash: boolean;
    useRightSquareBracket: boolean;
    useSpacingCircumflexAccent: boolean;
    useSpacingUnderscore: boolean;
    useSpacingGraveAccent: boolean;
    useLeftBrace: boolean;
    useVerticalBar: boolean;
    useRightBrace: boolean;
    useTildeAccent: boolean;

    hasChanged: boolean;
    randStrings: string[]
}